requirements.txt
flask
langchain
faiss-cpu
sentence-transformers
llama-cpp-python
langchain-community
